add auth to the appointments creation routes, also improved structure creating new files like app.ts, router.ts and server.ts
