+++ src/server.ts
@@ -1,0 +1,7 @@
+import { app } from "./app";
+
+const PORT = process.env.PORT || 3000;
+
+app.listen(PORT, () => {
+  console.log(`Server running ${PORT}`);
+});
